Add CLI, system tests, and full getting-started README
- cli/main.py: argparse entry point with --symbol, --start, --end,
  --capital, --entry-pct, --exit-pct, --stop-loss, --output-dir, --no-cache
- __main__.py: enables `python -m alphaa`
- System test: full pipeline with recorded RELIANCE.NS fixture data,
  verifying trades, metrics, CSV export, and chart generation
- README: installation, quick start, custom strategy example, project
  structure, development setup
- Fix type annotations throughout for mypy strict compliance

98 tests, all passing. mypy clean. ruff clean.

+++ alphaa/reporting/metrics.py
@@ -121,7 +121,7 @@ def _compute_sharpe(equity: list, risk_free_rate: float = 0.0) -> float:  # type
         return 0.0
 
     daily_sharpe = (mean_return - risk_free_rate / 252) / std_return
-    return daily_sharpe * math.sqrt(252)
+    return float(daily_sharpe * math.sqrt(252))
 
 
 def _empty_metrics(total_trades: int, benchmark: list) -> BacktestMetrics:  # type: ignore[type-arg]
